Split tests class into 2 to be able to handle the export use cases
